Server Manager: Changed the default behaviour of server manager to upgrade to "3.13.0-34"
closes-bug: #1387895
Fix the manifest so that we dont do anything for other flavors.

+++ contrail/manifests/contrail_common.pp
@@ -31,20 +31,26 @@ define upgrade-kernel($contrail_kernel_version) {
     $headers = "linux-headers-${contrail_kernel_version}"
     $headers_generic = "linux-headers-${contrail_kernel_version}-generic"
     $image = "linux-image-${contrail_kernel_version}"
-    package { 'apparmor' : ensure => '2.7.102-0ubuntu3.10',}
-    ->
-    package { $headers : ensure => present, }
-    ->
-    package { $headers_generic : ensure => present, }
-    ->
-    package { $image : ensure => present, }
-    ->
-    exec { "upgrade-kernel-reboot":
+
+    if ($operatingsystem == "Ubuntu")
+    {
+        package { 'apparmor' : ensure => '2.7.102-0ubuntu3.10',}
+        ->
+        package { $headers : ensure => present, }
+        ->
+        package { $headers_generic : ensure => present, }
+        ->
+        package { $image : ensure => present, }
+        ->
+        exec { "upgrade-kernel-reboot":
         command => "echo upgrade-kernel-reboot >> /etc/contrail/contrail_common_exec.out && reboot ",
-        provider => shell,
-        logoutput => "true",
-	unless => ["grep -qx upgrade-kernel-reboot /etc/contrail/contrail_common_exec.out"]
-    }
+            provider => shell,
+            logoutput => "true",
+	    unless => ["grep -qx upgrade-kernel-reboot /etc/contrail/contrail_common_exec.out"]
+    	}
+     } else {
+        #TODO for other flavours do nothing
+     }
 }
 
 #end of upgrade-kernel
